Fix: Only one character per name.
When trying to create a character, the system checks if a character with the same name already exists. If it exists,  that character is update with the new information. If not, a new character is created.  - NOTE: This should be changed when we include the field larp in the character class. There should be only one character with the same name per larp.